Change general factories into more concrete JXG-specific factories - change color of movable point on angle classification page
